Yam 18.6 00:04 add genre and recent added functions , set of list books return , book utils added
