+++ app.py
@@ -5,6 +5,7 @@ from auth.auth import auth
 from books.books import books
 from ai_utils.memoryManager import initialize_app
 from ai_utils.ai_routes import ai_story
+from werkzeug.serving import WSGIRequestHandler
 app = Flask(__name__)
 app.config["JWT_SECRET_KEY"] = Config.SECRET_KEY
 
@@ -20,9 +21,7 @@ if __name__ == "__main__":
 
     #התקנה של פיירבאס רק להרצה , ברגע שתוריד את הפייבאס תוריד את השורה הזו 
     initialize_app()
-
-    # for rule in app.url_map.iter_rules():
-    #     print(rule)
+    WSGIRequestHandler.protocol_version = "HTTP/1.1"
 
     app.run(host="0.0.0.0", port=5000, debug=True)
